# internal/libs/protoio: optimize MarshalDelimited by plain byteslice allocations+sync.Pool