@odeke-em 2021-12-10
Cool, I've performed the rebase! Thank you very much for the reviews and the time @marbar3778 @creachadair @tychoish, much appreciated and great work on maintaining Tendermint!